Make input and divide py file.
