+++ auto-drawing-tool/auto_drawing_tool.py
@@ -1,18 +1,20 @@
 import bpy
-import math
-import mathutils
+
+from .divide_frame import sortObjectAlongCurve
+from .divide_frame import divideFrame
+
 
 # Main function.
 def autoDraw(frame_range=None, basic=None, bl_render=None,
              material=None, world=None, modifier=None,
              sort=None, freestyle_preset=None, line_thick=None,
-             divide_frame=None, along_curve=None, sort_curve=None):
+             divide_frame=None, sort_along_curve=None):
     
     # Loop through selected objects.
     selected_objects = bpy.context.selected_objects
     
-    # curveの0点と選択したオブジェクトの距離ごとにソートして、それに従って、フレーム範囲を分割して順に描きたい---------。
-    if along_curve == True:
+    # curveに近いオブジェクトから先に描く：
+    if divide_frame == 'ALONG_CURVE':
         if bpy.context.active_object.type == 'CURVE':
             active_curve = bpy.context.active_object
             bezier_points = active_curve.data.splines[0].bezier_points
@@ -24,10 +26,14 @@ def autoDraw(frame_range=None, basic=None, bl_render=None,
             selected_objects = sorted_objects
             
             # curveに近いvertexの順にソート：
-            for obj_info in sorted_objects_info:
-                sortVertsAlongCurve(obj=obj_info['object'], location=obj_info['coordinate'])
+            '''
+            if sort_along_curve == True:
+                for obj_info in sorted_objects_info:
+                    curveSort(obj=obj_info['object'], location=obj_info['coordinate'])
+            '''
     
-    if divide_frame == True:
+    # フレームをオブジェクトで分配：
+    if divide_frame in ['ALONG_CURVE', 'SIMPLE_DIVIDE']:
         divided_frame_step = divideFrame(objects=selected_objects, frame_range=frame_range)
     
     for i, selected_object in enumerate(selected_objects):
@@ -37,10 +43,9 @@ def autoDraw(frame_range=None, basic=None, bl_render=None,
     
         # Only work for mesh, curve, or text.
         if bpy.context.object.type in ['MESH','CURVE','FONT']:
-            if divide_frame == True:
+            if divide_frame in ['ALONG_CURVE', 'SIMPLE_DIVIDE']:
                 frame_range[0] = divided_frame_step * i
                 frame_range[1] = divided_frame_step * (i+1)
-                print(frame_range)
 
             # Turn on/off each step--------------------
             if basic == True:
@@ -76,80 +81,6 @@ def autoDraw(frame_range=None, basic=None, bl_render=None,
             if line_thick == None:
                 line_thick = 2
             bpy.context.scene.render.line_thickness = line_thick
-
-# kd-treeのベース関数：
-def kdFind(data, point):
-    # サイズを入力：
-    size = len(data)
-    kd = mathutils.kdtree.KDTree(size)
-    
-    for i, d in enumerate(data):
-        kd.insert(d, i)
-    # 分割したデータ数を均等にするために、かならずfindの前に必要：
-    kd.balance()
-    # ある点に距離が近いものを挙げる：
-    co, index, dist = kd.find(point)
-    return [co, index, dist]
-
-# オブジェクトのvertexをworld座標に変換してkd-tree実行：
-def kdFindNearestPoint(obj, point=(0,0,0)):
-    world_verts = [obj.matrix_world * v.co for v in obj.data.vertices]
-    result = kdFind(world_verts, point)
-    # 指標のために、オブジェクトをリストの最後に加えてreturn。
-    result.append(obj)
-    # return [座標, vertex番号, 距離,　オブジェクト]：
-    return result
-
-def addNearestObject(objects, point):
-    nearest = {'distance': 10000, 'data':[], 'object': None}
-    for obj in objects:
-        # curveの0点に近いオブジェクトのvertexを挙げる。[座標, vertex番号, 距離,　オブジェクト]が出る。
-        near_result = kdFindNearestPoint(obj=obj, point=point)
-        if near_result[2] < nearest['distance']:
-            nearest['distance'] = near_result[2]
-            nearest['object'] = near_result[3]
-            nearest['coordinate'] = near_result[0]
-            nearest['vertex_index'] = near_result[1]
-    # return {'distance': n, 'object': obj, 'coordinate': 座標, 'vertex_index': vertex番号}
-    return nearest
-
-def sortObjectAlongCurve(objects, bezier_points):
-        # curveの各pointに、いくつのオブジェクトを割り当てるかのリスト生成：
-        step = len(bezier_points) / len(objects)
-        bezier_points_index = [math.floor(step*i) for i in range(len(objects))]
-        
-        sorted_objects = []
-        sorted_objects_info = []
-        
-        # curveのポイントの分だけ、最も近いオブジェクトをリストに入れる：
-        for i in bezier_points_index:
-            point_co = bezier_points[i].co
-            
-            # active_curveの0点と最も近いvertexを持つオブジェクト：
-            nearest_object = addNearestObject(objects=objects, point=point_co)
-            
-            # ターゲットリストから削除する：
-            objects.remove(nearest_object['object'])
-            # 新しいオブジェクトリストに近い方から追加する：
-            sorted_objects.append(nearest_object['object'])
-            sorted_objects_info.append(nearest_object)
-            
-        return sorted_objects, sorted_objects_info
-
-def sortVertsAlongCurve(obj, location):
-    bpy.ops.object.select_all(action='DESELECT')
-    obj.select = True
-    bpy.context.scene.objects.active = obj
-    # カーソルをcurveに近いvertexに変えて、CURSOR_DISTANCEでソート：
-    bpy.context.scene.cursor_location = location
-    changeSort(sort_type='CURSOR_DISTANCE')
-    bpy.ops.object.select_all(action='DESELECT')
-
-# 選択オブジェクトの数でbuildのフレームを分割：
-def divideFrame(objects, frame_range):
-    frame_duration = (frame_range[1] + 1) - frame_range[0]
-    frame_step = math.floor(frame_duration / len(objects))
-    return frame_step
 
 # Activate build modifier and freestyle.
 def addBuildFreestyle(frame_range):
@@ -222,6 +153,15 @@ def cameraViewSort():
     bpy.context.scene.cursor_location = cam.location
     changeSort(sort_type='CURSOR_DISTANCE')
     bpy.context.scene.cursor_location = [0,0,0]
+
+def curveSort(obj, location):
+    bpy.ops.object.select_all(action='DESELECT')
+    obj.select = True
+    bpy.context.scene.objects.active = obj
+    # カーソルをcurveに近いvertexに変えて、CURSOR_DISTANCEでソート：
+    bpy.context.scene.cursor_location = location
+    changeSort(sort_type='CURSOR_DISTANCE')
+    bpy.ops.object.select_all(action='DESELECT')
 
 # Set a freestyle preset.
 def setFreestylePreset(freestyle_preset):
